fix md5 check option

+++ main.py
@@ -17,9 +17,9 @@ from utils.md5 import verify_md5
 @click.option('--username', '-u', type=str, default='anonymous')
 @click.option('--password', '-p', type=str, default='')
 @click.option('--concurrency', type=int, default=3, help='最大并发下载数')
+@click.option('--md5', is_flag=True, help='是否进行md5校验', default=False)
 @click.option('--prefix', type=str, help='前缀筛选')
 @click.option('--suffix', type=str, help='后缀筛选')
-@click.option('--md5', is_flag=True, help='是否进行md5校验', default=False)
 def download(
     url: str,
     local_path: str,
@@ -27,6 +27,7 @@ def download(
     username: str,
     password: str,
     concurrency: int,
+    md5: bool,
     prefix: Optional[str] = None,
     suffix: Optional[str] = None,
 ):
@@ -35,13 +36,15 @@ def download(
         exit(0)
 
     parsed = urlparse(str(url))
-    host, path = parsed.netloc, parsed.path
+    host, remote_path = parsed.netloc, parsed.path
 
     with FTPClient(host=host, username=username, password=password) as ftp:
-        files = ftp.list_files(path)
+        files = ftp.list_files(remote_path)
 
     exist_files = []
     download_dir = Path(local_path)
+    download_dir.mkdir(parents=True, exist_ok=True)
+
     for file_path in download_dir.iterdir():
         if file_path.is_file():
             if file_path.stat().st_size == 0:
@@ -59,9 +62,9 @@ def download(
 
     async def download_func():
         await download_http_files(
-            base_url='https://ftp.ncbi.nlm.nih.gov',
-            remote_path='/pubmed/baseline',
-            file_list=list(missed_files),
+            base_url=f'https://{host}',
+            remote_path=remote_path,
+            file_list=list(missed_files)[:10],
             local_dir=local_path,
             max_concurrency=concurrency,
         )
@@ -70,20 +73,21 @@ def download(
         asyncio.run(download_func())
 
     # 校验 MD5
-    gz_list = list(download_dir.glob('*.gz'))
-    fail_count = 0
-    for gz_file in tqdm(gz_list, total=len(gz_list), desc='MD5校验'):
-        md5_file = download_dir / f'{gz_file.name}.md5'
-        if not verify_md5(gz_file, md5_file):
-            gz_file.unlink(missing_ok=True)
-            md5_file.unlink(missing_ok=True)
-            fail_count += 1
+    if md5:
+        gz_list = list(download_dir.glob('*.gz'))
+        fail_count = 0
+        for gz_file in tqdm(gz_list, total=len(gz_list), desc='MD5校验'):
+            md5_file = download_dir / f'{gz_file.name}.md5'
+            if not verify_md5(gz_file, md5_file):
+                gz_file.unlink(missing_ok=True)
+                md5_file.unlink(missing_ok=True)
+                fail_count += 1
 
-    if fail_count > 0:
-        logger.error(f'有 {fail_count} 个文件MD5校验失败，已删除。请重新运行此命令以下载')
-        exit(0)
-    else:
-        logger.info('MD5校验通过')
+        if fail_count > 0:
+            logger.error(f'有 {fail_count} 个文件MD5校验失败，已删除。请重新运行此命令以下载')
+            exit(0)
+        else:
+            logger.info('MD5校验通过')
 
 
 if __name__ == '__main__':
